backend/requirements.txt
fastapi==0.135.1
google-genai==1.67.0
pillow>=11.0.0
python-dotenv>=1.1.0
python-multipart>=0.0.20
uvicorn[standard]>=0.35.0
